Allow the default volume device name to be changed, and volume auto-create default value to be configured.
Requires the following local_settings:

REST_API_REQUIRED_SETTINGS = ['LAUNCH_INSTANCE_DEFAULTS']

LAUNCH_INSTANCE_DEFAULTS = {
    'vol_device_name': 'sda',
    'vol_create': True
}

When an image has the attribute set to use "hw_scsi_model=virtio-scsi", then the default device name should be sda.

If the default device name remains 'vda', then it is not possible to add a second disk.

This is because the LUN ID assigned to the second disk will conflict with the first if the device names are different.

Setting the default volume device name is required when